hCalc virker - har ikke kommentar endnu

+++ printMap.c
@@ -4,10 +4,10 @@
 
 #include "header.h"
 
-void printMap(int* map, const int mapSize) { //We are printing the map
+void printMap(double* map, const int mapSize) { //We are printing the map
     for (int i = 0; i < mapSize; i++){
         for (int j = 0; j < mapSize; j++){
-            printf(" %d ", map[i * mapSize + j]);
+            printf(" %0.0lf ", map[i * mapSize + j]);
         }
         printf("\n");
     }
